[Bug 445] Ancient faxes get re-sent, and to the wrong people!!!
Nothing to do with ancient faxes, but with faxq blocking until the exit
of a child (notably bin/wedged, and bin/notify).  "Files" being left in
the docq for "recycling" is one of the side-effects of the blocking.

This patch removes the unnecessary blocking in cases where faxq already
has the framework to handle children properly, and prevents this situation
from ever arising.