The Waiting for jobcontrol wasn't quite robust enough
There are times when we have to wait, but there are times when we really don't
want to wait.  The previous move of the wait into the schedule was an attempt
to remove the wait from some parts, but went too far.  This patch makes the
waiting more robust...

See Bug 745