This patch fixes a race in the batching/scheduling code of faxq.
This race has always existed, but now that JobControl can make a run through
the runScheduler() take a long time, the race is highly visible.

runScheduler() is controled by faxQueueApp's schedTimeout.  Whenever a time
expires, runScheduler() is called.  The SchedTimeout::start contains logic to
try and "throttle" the runScheduler() calls, and not have it run concurrently,
but if runScheduler() ever takes more than 1 second, it can.

This patch puts a guard in faxQueueApp so it won't schedule concurrently (which
causes lots of badness in runq,sleepq, etc), and makes the
SchedTimeout::timeoutExpired() aware of this predicament so it keeps delaying
the runScheduler() if it's currently running.