This patch fixes a race in the batching/scheduling code of faxq.
The race has always been present, but now that JobControl greatly extends the
time spend operating on the sleepq, the race is highly visible, not requiring
an exact signal during the sleepq processing.

sleepq get's job's added and removed from it asynchronously.  It was never mant
to be "run" through in order.  It was originally only added to or removed from
in the timerExpire() routines in a manner that could not be interrupted.

But with the addition of batching scanning the sleepq for jobs to batch, it now
processes the sleepq in a manner that can be interupted.  JobControl
exasperates this window.

If a "sleeping" job has something happen to it (like it's TTS timer expires),
it will asynchronously go to ready (or some other state), and be put on some
other list.  Once it's on the other list, it's next field can never point to
the sleepiter.head starting position, causing an infinite loop
(sleepiter.notDone() will never be true.