Bug 783
Some fax machines, notably Brothers, don't like how we will retransmit
a single frame in duplicate.  This causes them to truncate the image at
that point.  However, some receivers will always miss the first frame
transmitted, and so we must eventually send single frames through in
duplicate to get them through in those cases.

So the first time we retransmit a single frame we do it alone - this
should catch nearly all of the Brother-like situations.  Then, if we
have to retransmit that single frame yet again we do it then in duplicate.