Move to a single runq in faxQueueApp
The idea of multiple runqs was to split up the job priorities.  This
shortens the time to do the sorted insert into the runq.  But as it
happens, most people don't use priorities, and the "large queues" which are
the ones that the sort has the biggest impact on are almost always
submitted at the same priority anyways.

Moving to a single runq, though theoretically a bit slower, in reality
isn't a performance hit.

And it simplifies the code.