Make the runq contain DestInfo instead of Job objects
This make the DestInfo readQ the primary source of "ready" jobs.  Any job
that is ready is in there.  The runq is a sorted list of DestInfo pointers
instead of Job pointers.  The readyQ inside each DestInfo is a sorted list
of the jobs (based on pri, just like runq used to be), and the DestInfo
pointers are sorted in runq based on their first job's sorting priority.

This means that the next job to send is *always* the first job in the first
DestInfo in the runq with jobs.

The runq can contain DestInfo objects that don't have any jobs on their
readyQ.  This happens they job is sending.  But since we know that the
number of these objects is bounded by the number of devices, which is tiny
in comparison.