Make a Batch the unit of work, instead of a Job.
Instead of handling a "Job" through the prepare/send/done machinery, we
handle a "Batch".   The Batch can add jobs to itself from those that are in
the DestInfo's ready queue.  It adds these jobs as it finishes preparing
previous jobs.

As a batch works, it's very easy to see progress go through the Jobs.
Because we can see that, we can act on jobs as they finish, moving them
through the queue much more quickly.