Check JobControlInfo before batching jobs together
This introduced the ability to make sure that jobs with different
"JobControl" options don't get batched together.

The initial compatibility is limited to the "args" that JobControlInfo
collects being the identical.  This could possibly be refined further but
most of the JobControlInfo members don't affect the actual faxsend
invocation so don't really matter for batching.