Short-circuit queue scan when no devices are ready
If no modem devices are available, then there is no point looking
for a job on the queue to send.

Because jobs are now always "Ready" and on the DestInfo list we
don't need to scan the entire queue to find them either.

Job timeouts/expiry are all handled by timers, so we don't need to
scan the queue to perform any of that work.