faxq: Don't release modem from the middle of a batch
Fix a bug where the modem is "released" in the middle of the batch.  If the
preparation failes, and there are other jobs in the batch (before or after), we
need to take care that delayJob/deleteJob don't release the modem, because we
need it for the rest of the jobs.