faxGetty: fix some EXTERN GETTY bugs
In the work on egetty, it changed from modemAnswerCallCmd to
modemAnswerCall.  Unfortunately, this broke the use of the
ModemAnswer*BeginCmd config options which used to be used, and chagne to
the ModeAnswer*Cmd ones (which the answerCall() routines use.  In
addition, the answerCall() routine actively *disallows* setting any of
those to blank, making it impossible to have nothing be sent to the modem
for any of these cases.  It would blindly send ATA if any of them were set
to blank.

This fixes some bugs in the extern handling, updates the getty stuff to
allow CallID passing as arguments, and reverts back to useing
modemAnswerCallCmd() instead of modemAnswerCall().