FaxRequest: Status codes for jobs
This patch introduces a Status object, and uses it in FaxRequest instead
of the fxStr notice to make sure we always have a code associated with any
error/status change in a job.

It introduces a new JPARM STATUSCODE command in the clinet-server protocol
so clients can retreive the "code" of the status as well as the status
string.

Status codes are documented in faxd/STATUS.txt