Resuffle a bit to make proper release notes easier.
The html/ directory is old and crufty, and already archived:
        http://www.hylafax.org/site1/

The README is now more simple, and points to the version specific
release notes, which contains information about new features, possible
incompatibilities, and changes.

Release notes can be accumulated on a per-release basis, and should
remain in the tree for past versions too.