Correct trigger spec parsing in faxq
When parsing triggers, cp was actually kept to point a character ahead
of the next character.  This meant that at the top of the while loop, it
was past the end of the the string it was supposed to be parsing.

With a busy faxq fifo, "past the end of the string" is pointing at the
next command faxq would normally process.  Trigger::parse() was trying
to parse that command, and erroring, because it is not a trigger spec.

This makes sure that the trigger spec is parsed correctly, and will not
move on trying to parse "other" memory.