faxmail: use mimeconverters to preprocess/discard mime parts
This adds back in the mimeconverter feature.  It is handled slightly
differently yet - individual parts can be processed through mimeconverter
scripts, but their output is still independant and submitted as seperate
documents to HylaFAX.

The mimeconverter is give the filename as it's only argument, and all
output is collected as a new document.  The converter can choose to do
whtever it wants with the document, including:

1) "filter" it, and different output that faxmail can submit as a document
2) Suppress it (by not producing any output) so it's not submitted
3) change/subsitute it (by producing any output it wants)

The output of the miemconverter is gathered and submitted as a document
with the Job, as long as it's not empty.  The output is *still* a separate
document, and handled through the normal typerules mechanism, so you can
do stuff like "sanitize" html, or "watermark" images, or "choose"
multipart/alternative preferences and still let the normal HylaFAX client
typerules handle any possible needed conversions to PS/PDF/TIFF for the
server submission.