hfaxd lockJob() - don't blindly O_CREAT job we're trying to lock
It happens that in some cases, we are only tryhing to lock a job to
updateJobFromDisk(), and this can happen *before* the qfile was renamed (to the
doneq), but after it was stated (successfully).  Opening wiht O_CREAT here
means that we end up "creating" these 0 length files when we shouldn't have
been able to lock it.

So we carefully create the file if we might need it, the lockJob can only
open it O_RDWR