i18n: Build infrastructure
This allows us to "build" .po and .pot files for HylaFAX.

It builds 3 sets:
        libhylafax              Strings used in libhylafax
        hylafax-client          Strings used in hylafax client components
        hylafax-server          Strings used in hylafax server components

All are merged to one master "hylafax.pot" for translating.  The 3
separate catalogs translations are merged out of the master
hylafax.pot/$LANG.po to build the .mo for installation.

Makefile hackery makes me which we only supported "real" makes...