hfaxd sorting based on the *SortFormat variables
This introduces the following new protocol commands
        MDMSORTFMT <format>
        JOBSORTFMT <format>
        RCVSORTFMT <format>
        FILESORTFMT <format>

These commands take the exact same format specification as the related:
        MDMFMT <format>
        JOBFMT <format>
        RCVFMT <format>
        FILEFMT <format>

If one of them is set, when the server generates the listing for the
status, sendq/doneq, or recvq, it will sort the list using the specified
format before sending it to the client.  They all default to nothing,
which allows hfaxd to send output directly to the client with no sorting.
The administrator can change the default value for these in hfaxd.conf.

It also introduces a new hfaxd boolean config option:
        AllowSortFormat
This defaults to true to allow the client to specify the sorting
format, and the administrator can set it to false to not let the client
request a different sorting format.  This allows the adminstator to make
sure the server doesn't sort at all, or to only use the sorting set in
hfaxd.conf.