faxmail: application/postscript attachments were sill going inline
The move to handle attachments as seperate documents and not try and munge
them all into a single postscript document forgot the
application/postscript case.

This changes it so application/postscript are handled as seperate
documents, just like everything else.  It adds a special mime type that
will allow postscript to be added inline:
        application/x-faxmail-inline-postscript