hfaxd: Allow others (same uid) to operate on jobs
HylaFAX generally considers UID mathing to be "equivilent" in terms of
logged in uses.  This extends the permission checking in hfaxd's sendq
file handling so the admin can set JobPermission such that users with the
same uid have "write" permission, allows them to modify jobs just like
the jobs owner.