hfaxd - IPv6 (AF_INET6) socket support
Now we can accept both INET and INET6 sockets...  We used to be able to do
that with xinetd/etc, but we would handle an INET6 socket like an INET
socket.  Now we can do both natively, or with an inetd helper, and handle
both properly as an INET/INET6 sockets.