Out of tree build - message .mo catalogs
In the conversion to not use inferred rules (supporting old makes), I
dropped the ability to build out-of-tree from my initial NLS work.  With
an inferred rule, make handle the path stuff nicely for us.

Here we check if $CATALOG/messages.pot exists, and if not, copy it from
$SRCDIR.  The only reason for it to not exist is that we're not building
in SRCDIR.

Note that to build the .mo files, you either need to have the
messages.pot for the various catalog.  Either "build" them
(make update-po), or copy them from SRCDIR.