Rework NLS slightly for old SunOS
We need to be careful not to do any real gettext defines.  We tried to
avoid touching any real gettext stuff throught the code for fears of
portabilty, this just finishes this up so our "code" is completely gettext
free, and we don't need to do any gettext defines that can conflict with
system headers.