Do not warn about one of the Fontpath directories not existing
Fontpath usually contain multiple directories with one or more
not existing. Having faxsetup warn the user about this directory
not existing and having the possibility of causing problems just
confuse users. There is already a warning if no .afm file is
found in any of those directories. That should be enough.