Reworked how faxsetup looks for Fontmap
If a previous faxsetup had added Fontmap/Fontpath lines in
hyla.conf and the hardcoded Fontmap path is the same as the
installed Ghostscript search path, the previously added
Fontmap/Fontpath lines were left untouched even if they
were not valid anymore.

This patch now looks for a valid combination to set in hyla.conf
and will remove unneeded lines if necessary. It will even set
a default font (using a font filename) if the default "Courier"
font (or any font specified in hyla.conf) is not found using
the Fontmap files.