DialRules - recursively allow dialrules to reference others
This allows rules to reference other rules.  The names are looked up
at runtime, not parse time, so they can reference rules not yet
defined.  The expected usage would be to reference a "canonicalization" rule
(either CanonicalNumber directly, or through some indepenant rule) in
multiple places without needing repeats.

The replace format is of the form be \Name(...)

Example dialrules
    ! rules to canonoicalize numbers to +IDD format
    CanonicalNumber := [
	....
    ]
    DialString := [
	.*			= \CanonicalNumber(&)	! Get the canonical
	^[+]${Country}${Area}	= 			! Local call
	^[+]${Country}		= ${LDPrefix}		! domestic long distance
	^[+]			= ${IDDPrefix}		! International
    ]

There is a recursion depth limit in this, just to avoid infinute rule recursion.